Create Function and lists Js
In this course, I'm going to learn how to create functions and use JavaScript to change elements on the screen. I'll improve my programming logic with a hands-on project, as well as learn how to use GitHub and Vercel to publish my projects. I'll also figure out how to work with lists, an important structure in programming, and manipulate words and numbers dynamically.